decode_argmax2 + sifter event protocol: float -> double (all CPU-side; no fp64 on GPU)
- DedispersionPlan::decode_argmax2(): double outputs/internals (host-only code)
- decode_argmax2_batch: returns float64 numpy arrays
- frb_sifter.proto: all float fields -> double (FrbEvent + coarsegrain_snr);
  PROTOCOL_VERSION_CURRENT 3 -> 4; refreshed stale placeholder comment
- FrbSifterEvents/create_events(): float64 dtypes end-to-end

Note: the sifter side (chord-frb-sifter repo) needs the matching proto edit,
since a version-3 sifter now rejects the config handshake.

+++ include/pirate/DedispersionPlan.hpp
@@ -120,8 +120,8 @@ struct DedispersionPlan
     
     void decode_argmax2(
         long itree, long fmin, long fmax, long tlo, long thi, long p,
-        float &freq_lo_MHz, float &freq_hi_MHz, float &dm,
-        float &timestamp_samp, float &width_samp) const;
+        double &freq_lo_MHz, double &freq_hi_MHz, double &dm,
+        double &timestamp_samp, double &width_samp) const;
 
 
     // An "incomplete" DedispersionPlan does not initialize any of the "low-level data needed
